# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AirGo前后分离,多用户代理面板",
        "title": "AirGo",
        "contact": {
            "url": "https://github.com/pura-panel/airgo-panel"
        },
        "license": {
            "name": "GPL v3.0",
            "url": "https://github.com/pura-panel/airgo-panel/blob/main/LICENSE"
        }
    },
    "paths": {
...
        "/api/admin/node/createx25519": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] node"
                ],
                "summary": "reality x25519",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/node/deleteNode": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] node"
                ],
                "summary": "删除节点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Node"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/node/getNodeList": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] node"
                ],
                "summary": "获取节点列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QueryParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/node/getNodeListWithTraffic": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] node"
                ],
                "summary": "获取节点列表，带流量信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QueryParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/node/getNodeServerStatus": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] node"
                ],
                "summary": "获取节点服务器状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/node/newNode": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] node"
                ],
                "summary": "新建节点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Node"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/node/newNodeShared": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] node"
                ],
                "summary": "新增共享节点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Node"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/node/nodeSort": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] node"
                ],
                "summary": "节点排序",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Node"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/node/parseUrl": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] node"
                ],
                "summary": "解析",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NodeSharedReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/node/updateNode": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] node"
                ],
                "summary": "更新节点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Node"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Access": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "route": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Coupon": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "discount_rate": {
                    "type": "number"
                },
                "expired_at": {
                    "type": "string"
                },
                "goods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Goods"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "min_amount": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.FieldParamsItem": {
            "type": "object",
            "properties": {
                "condition": {
                    "description": "= \u003e \u003c \u003c\u003e like",
                    "type": "string"
                },
                "condition_value": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "field_type": {
                    "type": "string"
                },
                "operator": {
                    "description": "AND OR",
                    "type": "string"
                }
            }
        },
...
        "model.Goods": {
            "type": "object",
            "properties": {
                "coupon": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Coupon"
                    }
                },
                "cover_image": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deliver_text": {
                    "type": "string"
                },
                "deliver_type": {
                    "type": "string"
                },
                "des": {
                    "type": "string"
                },
                "enable_traffic_reset": {
                    "type": "boolean"
                },
                "goods_order": {
                    "description": "基础参数",
                    "type": "integer"
                },
                "goods_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_renew": {
                    "type": "boolean"
                },
                "is_sale": {
                    "type": "boolean"
                },
                "is_show": {
                    "type": "boolean"
                },
                "node_connector": {
                    "type": "integer"
                },
                "node_speed_limit": {
                    "type": "integer"
                },
                "nodes": {
                    "description": "关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Node"
                    }
                },
                "price": {
                    "description": "todo 前端修改",
                    "type": "string"
                },
                "price_12_month": {
                    "type": "string"
                },
                "price_3_month": {
                    "description": "订阅参数",
                    "type": "string"
                },
                "price_6_month": {
                    "type": "string"
                },
                "price_unlimited_duration": {
                    "type": "string"
                },
                "quota": {
                    "type": "integer"
                },
                "recharge_amount": {
                    "description": "充值参数",
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "subject": {
                    "type": "string"
                },
                "total_bandwidth": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Node": {
            "type": "object",
            "properties": {
                "access": {
                    "description": "访问控制",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Access"
                    }
                },
                "address": {
                    "type": "string"
                },
                "aid": {
                    "type": "integer"
                },
                "allowInsecure": {
                    "type": "boolean"
                },
                "alpn": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "dest": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "encryption": {
                    "type": "string"
                },
                "flow": {
                    "type": "string"
                },
                "fp": {
                    "type": "string"
                },
                "goods": {
                    "description": "关联参数",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Goods"
                    }
                },
                "host": {
                    "type": "string"
                },
                "hy_down_mbps": {
                    "type": "integer"
                },
                "hy_obfs": {
                    "type": "string"
                },
                "hy_obfs_password": {
                    "type": "string"
                },
                "hy_ports": {
                    "type": "string"
                },
                "hy_up_mbps": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "mode": {
                    "type": "string"
                },
                "network": {
                    "type": "string"
                },
                "node_order": {
                    "type": "integer"
                },
                "node_speed_limit": {
                    "type": "integer"
                },
                "node_type": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "pbk": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "private_key": {
                    "type": "string"
                },
                "protocol": {
                    "description": "一些协议参数",
                    "type": "string"
                },
                "remarks": {
                    "description": "基础参数",
                    "type": "string"
                },
                "scy": {
                    "type": "string"
                },
                "security": {
                    "type": "string"
                },
                "server_key": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "sid": {
                    "type": "string"
                },
                "sni": {
                    "type": "string"
                },
                "spx": {
                    "type": "string"
                },
                "total_down": {
                    "description": "Byte",
                    "type": "integer"
                },
                "total_up": {
                    "description": "已用上行/已用下行，统计节点流量时使用",
                    "type": "integer"
                },
                "traffic_rate": {
                    "type": "number"
                },
                "transfer_address": {
                    "description": "中转参数",
                    "type": "string"
                },
                "transfer_node_id": {
                    "type": "integer"
                },
                "transfer_port": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "description": "共享节点需要的uuid;订阅下发是实际的用户uuid",
                    "type": "string"
                },
                "v": {
                    "type": "string"
                }
            }
        },
        "model.NodeSharedReq": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.Pagination": {
            "type": "object",
            "properties": {
                "order_by": {
                    "type": "string"
                },
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
...
        "model.QueryParams": {
            "type": "object",
            "required": [
                "table_name"
            ],
            "properties": {
                "field_params_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FieldParamsItem"
                    }
                },
                "pagination": {
                    "description": "分页参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Pagination"
                        }
                    ]
                },
                "table_name": {
                    "type": "string"
                }
            }
        },
...
        "response.ResponseStruct": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}